refactor(core): move AI workflows into workflow directories
