+++ packages/core/src/ai-model/workflows/inspect/index.ts
@@ -4,4 +4,4 @@ export {
   AiLocateElement,
   AiLocateSection,
   buildSearchAreaConfig,
-} from '../../inspect';
+} from './inspect';
